feat: removes old Linode provider.
