+++ providers/dns/linode/linode.go
@@ -1,15 +1,19 @@
-// Package linode implements a DNS provider for solving the DNS-01 challenge using Linode DNS.
+// Package linode implements a DNS provider for solving the DNS-01 challenge using Linode DNS and Linode's APIv4
 package linode
 
 import (
+	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
+	"net/http"
 	"strings"
 	"time"
 
 	"github.com/go-acme/lego/v3/challenge/dns01"
 	"github.com/go-acme/lego/v3/platform/config/env"
-	"github.com/timewasted/linode/dns"
+	"github.com/linode/linodego"
+	"golang.org/x/oauth2"
 )
 
 const (
@@ -22,24 +26,30 @@ const (
 const (
 	envNamespace = "LINODE_"
 
-	EnvAPIKey = envNamespace + "API_KEY"
+	EnvToken = envNamespace + "TOKEN"
 
-	EnvTTL             = envNamespace + "TTL"
-	EnvPollingInterval = envNamespace + "POLLING_INTERVAL"
+	EnvTTL                = envNamespace + "TTL"
+	EnvPropagationTimeout = envNamespace + "PROPAGATION_TIMEOUT"
+	EnvPollingInterval    = envNamespace + "POLLING_INTERVAL"
+	EnvHTTPTimeout        = envNamespace + "HTTP_TIMEOUT"
 )
 
 // Config is used to configure the creation of the DNSProvider.
 type Config struct {
-	APIKey          string
-	PollingInterval time.Duration
-	TTL             int
+	Token              string
+	PropagationTimeout time.Duration
+	PollingInterval    time.Duration
+	TTL                int
+	HTTPTimeout        time.Duration
 }
 
 // NewDefaultConfig returns a default configuration for the DNSProvider.
 func NewDefaultConfig() *Config {
 	return &Config{
-		TTL:             env.GetOrDefaultInt(EnvTTL, minTTL),
-		PollingInterval: env.GetOrDefaultSecond(EnvPollingInterval, 15*time.Second),
+		TTL:                env.GetOrDefaultInt(EnvTTL, minTTL),
+		PropagationTimeout: env.GetOrDefaultSecond(EnvPropagationTimeout, 0),
+		PollingInterval:    env.GetOrDefaultSecond(EnvPollingInterval, 15*time.Second),
+		HTTPTimeout:        env.GetOrDefaultSecond(EnvHTTPTimeout, 0),
 	}
 }
 
@@ -51,19 +61,19 @@ type hostedZoneInfo struct {
 // DNSProvider implements the challenge.Provider interface.
 type DNSProvider struct {
 	config *Config
-	client *dns.DNS
+	client *linodego.Client
 }
 
 // NewDNSProvider returns a DNSProvider instance configured for Linode.
-// Credentials must be passed in the environment variable: LINODE_API_KEY.
+// Credentials must be passed in the environment variable: LINODE_TOKEN.
 func NewDNSProvider() (*DNSProvider, error) {
-	values, err := env.Get(EnvAPIKey)
+	values, err := env.Get(EnvToken)
 	if err != nil {
 		return nil, fmt.Errorf("linode: %w", err)
 	}
 
 	config := NewDefaultConfig()
-	config.APIKey = values[EnvAPIKey]
+	config.Token = values[EnvToken]
 
 	return NewDNSProviderConfig(config)
 }
@@ -74,35 +84,49 @@ func NewDNSProviderConfig(config *Config) (*DNSProvider, error) {
 		return nil, errors.New("linode: the configuration of the DNS provider is nil")
 	}
 
-	if len(config.APIKey) == 0 {
-		return nil, errors.New("linode: credentials missing")
+	if len(config.Token) == 0 {
+		return nil, errors.New("linode: Linode Access Token missing")
 	}
 
 	if config.TTL < minTTL {
 		return nil, fmt.Errorf("linode: invalid TTL, TTL (%d) must be greater than %d", config.TTL, minTTL)
 	}
 
+	tokenSource := oauth2.StaticTokenSource(&oauth2.Token{AccessToken: config.Token})
+	oauth2Client := &http.Client{
+		Timeout: config.HTTPTimeout,
+		Transport: &oauth2.Transport{
+			Source: tokenSource,
+		},
+	}
+
+	client := linodego.NewClient(oauth2Client)
+	client.SetUserAgent(fmt.Sprintf("lego-dns linodego/%s", linodego.Version))
+
 	return &DNSProvider{
 		config: config,
-		client: dns.New(config.APIKey),
+		client: &client,
 	}, nil
 }
 
 // Timeout returns the timeout and interval to use when checking for DNS
 // propagation.  Adjusting here to cope with spikes in propagation times.
-func (d *DNSProvider) Timeout() (timeout, interval time.Duration) {
-	// Since Linode only updates their zone files every X minutes, we need
-	// to figure out how many minutes we have to wait until we hit the next
-	// interval of X.  We then wait another couple of minutes, just to be
-	// safe.  Hopefully at some point during all of this, the record will
-	// have propagated throughout Linode's network.
-	minsRemaining := dnsUpdateFreqMins - (time.Now().Minute() % dnsUpdateFreqMins)
+func (d *DNSProvider) Timeout() (time.Duration, time.Duration) {
+	timeout := d.config.PropagationTimeout
+	if d.config.PropagationTimeout <= 0 {
+		// Since Linode only updates their zone files every X minutes, we need
+		// to figure out how many minutes we have to wait until we hit the next
+		// interval of X.  We then wait another couple of minutes, just to be
+		// safe.  Hopefully at some point during all of this, the record will
+		// have propagated throughout Linode's network.
+		minsRemaining := dnsUpdateFreqMins - (time.Now().Minute() % dnsUpdateFreqMins)
 
-	timeout = (time.Duration(minsRemaining) * time.Minute) +
-		(minTTL * time.Second) +
-		(dnsUpdateFudgeSecs * time.Second)
-	interval = d.config.PollingInterval
-	return
+		timeout = (time.Duration(minsRemaining) * time.Minute) +
+			(minTTL * time.Second) +
+			(dnsUpdateFudgeSecs * time.Second)
+	}
+
+	return timeout, d.config.PollingInterval
 }
 
 // Present creates a TXT record using the specified parameters.
@@ -113,39 +137,40 @@ func (d *DNSProvider) Present(domain, token, keyAuth string) error {
 		return err
 	}
 
-	if _, err = d.client.CreateDomainResourceTXT(zone.domainID, dns01.UnFqdn(fqdn), value, d.config.TTL); err != nil {
-		return err
+	createOpts := linodego.DomainRecordCreateOptions{
+		Name:   dns01.UnFqdn(fqdn),
+		Target: value,
+		TTLSec: d.config.TTL,
+		Type:   linodego.RecordTypeTXT,
 	}
 
-	return nil
+	_, err = d.client.CreateDomainRecord(context.Background(), zone.domainID, createOpts)
+	return err
 }
 
 // CleanUp removes the TXT record matching the specified parameters.
 func (d *DNSProvider) CleanUp(domain, token, keyAuth string) error {
 	fqdn, value := dns01.GetRecord(domain, keyAuth)
+
 	zone, err := d.getHostedZoneInfo(fqdn)
 	if err != nil {
 		return err
 	}
 
 	// Get all TXT records for the specified domain.
-	resources, err := d.client.GetResourcesByType(zone.domainID, "TXT")
+	listOpts := linodego.NewListOptions(0, "{\"type\":\"TXT\"}")
+	resources, err := d.client.ListDomainRecords(context.Background(), zone.domainID, listOpts)
 	if err != nil {
 		return err
 	}
 
 	// Remove the specified resource, if it exists.
 	for _, resource := range resources {
-		if resource.Name == zone.resourceName && resource.Target == value {
-			resp, err := d.client.DeleteDomainResource(resource.DomainID, resource.ResourceID)
-			if err != nil {
+		if (resource.Name == strings.TrimSuffix(fqdn, ".") || resource.Name == zone.resourceName) &&
+			resource.Target == value {
+			if err := d.client.DeleteDomainRecord(context.Background(), zone.domainID, resource.ID); err != nil {
 				return err
 			}
-
-			if resp.ResourceID != resource.ResourceID {
-				return errors.New("error deleting resource: resource IDs do not match")
-			}
-			break
 		}
 	}
 
@@ -159,16 +184,24 @@ func (d *DNSProvider) getHostedZoneInfo(fqdn string) (*hostedZoneInfo, error) {
 		return nil, err
 	}
 
-	resourceName := strings.TrimSuffix(fqdn, "."+authZone)
-
 	// Query the authority zone.
-	domain, err := d.client.GetDomain(dns01.UnFqdn(authZone))
+	data, err := json.Marshal(map[string]string{"domain": dns01.UnFqdn(authZone)})
 	if err != nil {
 		return nil, err
 	}
 
+	listOpts := linodego.NewListOptions(0, string(data))
+	domains, err := d.client.ListDomains(context.Background(), listOpts)
+	if err != nil {
+		return nil, err
+	}
+
+	if len(domains) == 0 {
+		return nil, errors.New("domain not found")
+	}
+
 	return &hostedZoneInfo{
-		domainID:     domain.DomainID,
-		resourceName: resourceName,
+		domainID:     domains[0].ID,
+		resourceName: strings.TrimSuffix(fqdn, "."+authZone),
 	}, nil
 }
